- 스태미나 회복 매커니즘 변경 - Character에 무게기능 추가 - ArmorData, BackpackData, RigData에 weight 추가 - Weight 정수로 변경 - 무게에 따른 이동력 변화 기능 추가

+++ Assets/Resources/Scrips/Data/BulletData.cs
@@ -18,7 +18,7 @@ public class BulletDataInfo
     [Tooltip("구경")] public float caliber;
     [Tooltip("펠릿 수")] public int pelletNum;
     [Tooltip("확산")] public int spread;
-    [Tooltip("무게")] public float weight;
+    [Tooltip("무게")] public int weight;
     [Space(5f)]
 
     [Tooltip("정확도")] public float MOA;
